+++ fs-util/src/directory_copier.rs
@@ -13,7 +13,10 @@ pub fn mirror_folder(from: &PathBuf, to: &Path) -> std::io::Result<()> {
         let target = to.join(target_name);
         if path.is_dir() {
             mirror_folder(&path, &target).unwrap()
-        } else if target.file_name().unwrap() != "manifest.acf" && !target.exists() {
+        } else if target.file_name().unwrap() != "manifest.acf"
+            && !target.exists()
+            && !target.is_symlink()
+        {
             windows_fs::symlink_file(path, target)?
         }
     }