+++ src/main/kotlin/omsi/OmsiInstaller.kt
@@ -86,21 +86,23 @@ private suspend fun doNativeCall(name: String, vararg parameters: String) =
         logger.debug { "Successfully finished native call." }
     }
 
-suspend fun activateInstallationSafe(path: Path, startSteam: Boolean = false, awaitSteamDeath: suspend () -> Boolean): Boolean {
+suspend fun activateInstallationSafe(instance: Instance, startSteam: Boolean = false, awaitSteamDeath: suspend () -> Boolean) {
     val omsi = getOmsiInstallPath()
-    if (omsi.isSymbolicLink() && omsi.readSymbolicLink() == path) {
+    if (omsi.isSymbolicLink() && omsi.readSymbolicLink() == instance.directory) {
         val currentManifest = omsi.parent(2) / "appmanifest_$OMSI_STEAM_ID.acf"
         if (currentManifest.exists()) {
-            logger.debug { "Backing up $currentManifest to current installation $path" }
-            currentManifest.copyTo(path / "manifest.acf", true)
-            return true
+            logger.debug { "Backing up $currentManifest to current installation ${instance.directory}" }
+            currentManifest.copyTo(instance.directory / "manifest.acf", true)
+            return
         }
     }
     val isSteamRunning = isSteamRunning()
 
     if (!isSteamRunning || awaitSteamDeath()) {
         doNativeCall("activate-omsi.exe", omsi.absolutePathString(), path.absolutePathString())
-
+        if (!(instance.directory / "Omsi.exe").exists()) {
+            reLinkOmsiExecutable(instance)
+        }
         if (isSteamRunning && startSteam) {
             withContext(Dispatchers.IO) {
                 runSteam()
@@ -109,11 +111,8 @@ suspend fun activateInstallationSafe(path: Path, startSteam: Boolean = false, aw
         return true
     }
 
-    return false
-}
-
-suspend fun activateAndStartInstallationSafe(path: Path, flags: List<LaunchFlag>, awaitSteamDeath: suspend () -> Boolean) {
-    if (activateInstallationSafe(path, awaitSteamDeath = awaitSteamDeath)) {
+suspend fun activateAndStartInstallationSafe(instance: Instance, flags: List<LaunchFlag>) {
+    if (activateInstallationSafe(instance, awaitSteamDeath = awaitSteamDeath)) {
         startOmsi(flags)
     }
 }